refactor(web): merge Server section into Performance in settings nav
Standalone sections for a single setting are hard to justify; combining
the heap memory control with thread cache size under one Performance
section keeps resource-tuning knobs co-located.

- Fold ServerSection content (heap memory) into PerformanceSection
- Delete ServerSection component
- Drop the "server" nav id; System group now has Performance + About
- Update settings-visual.spec.ts section list
- Add performance.threadCacheSize row to docs/settings/reference.md
