+++ apps/web/src/components/settings/settings-nav.ts
@@ -5,7 +5,6 @@ import { WorktreeSection } from "./sections/WorktreeSection";
 import { AppearanceSection } from "./sections/AppearanceSection";
 import { NotificationsSection } from "./sections/NotificationsSection";
 import { TerminalSection } from "./sections/TerminalSection";
-import { ServerSection } from "./sections/ServerSection";
 import { PerformanceSection } from "./sections/PerformanceSection";
 import { KeyboardShortcutsSection } from "./sections/KeyboardShortcutsSection";
 import { AboutSection } from "./sections/AboutSection";
@@ -19,7 +18,6 @@ export type SettingsSection =
   | "notifications"
   | "terminal"
   | "keyboard"
-  | "server"
   | "performance"
   | "about";
 
@@ -51,7 +49,6 @@ export const NAV_GROUPS: NavGroup[] = [
   {
     label: "System",
     items: [
-      { id: "server", label: "Server" },
       { id: "performance", label: "Performance" },
       { id: "about", label: "About" },
     ],
@@ -67,7 +64,6 @@ export const SECTION_MAP: Record<SettingsSection, ComponentType> = {
   notifications: NotificationsSection,
   terminal: TerminalSection,
   keyboard: KeyboardShortcutsSection,
-  server: ServerSection,
   performance: PerformanceSection,
   about: AboutSection,
 };
